fix : 6차... 달라진건없다...

+++ src/main/java/com/example/emotrak/dto/comment/CommentDetailResponseDto.java
@@ -36,7 +36,8 @@ public class CommentDetailResponseDto {
         this.nickname = comment.getUser().getNickname();
         this.likesCnt = (int) likesCnt;
         this.hasLike = userLikesCnt > 0;
-        this.hasReport = (int) reportCnt > 0;
+        this.hasReport = reportCnt > 0;
     }
+
 
 }
